+++ src/link/client-awareness/__tests__/ClientAwarenessLink.test.ts
@@ -1,31 +1,14 @@
 import fetchMock from "fetch-mock";
 
-import {
-  ApolloClient,
-  ApolloLink,
-  gql,
-  HttpLink,
-  InMemoryCache,
-  version,
-} from "@apollo/client";
-import {
-  BaseBatchHttpLink,
-  BatchHttpLink,
-} from "@apollo/client/link/batch-http";
+import { ApolloClient, gql, InMemoryCache, version } from "@apollo/client";
 import { ClientAwarenessLink } from "@apollo/client/link/client-awareness";
-import { BaseHttpLink } from "@apollo/client/link/http";
 import { MockSubscriptionLink } from "@apollo/client/testing";
-import { wait } from "@apollo/client/testing/internal";
 
 const query = gql`
   query {
     hello
   }
 `;
-const response = {
-  data: { hello: "world" },
-};
-const uri = "https://example.com/graphql";
 
 afterEach(() => fetchMock.reset());
 
@@ -239,230 +222,6 @@ describe("feature: client awareness", () => {
       });
     });
   });
-
-  describe("HttpLink integration", () => {
-    test("is part of `HttpLink`", () => {
-      fetchMock.postOnce(uri, response);
-      const client = new ApolloClient({
-        link: new HttpLink({
-          uri,
-        }),
-        cache: new InMemoryCache(),
-        clientAwareness: {
-          name: "test-client",
-          version: "1.0.0",
-        },
-      });
-
-      void client.query({ query });
-      const headers = fetchMock.lastCall()![1]?.headers;
-      expect(headers).toStrictEqual({
-        accept: "application/graphql-response+json,application/json;q=0.9",
-        "content-type": "application/json",
-        "apollographql-client-name": "test-client",
-        "apollographql-client-version": "1.0.0",
-      });
-    });
-
-    test("is not part of `BaseHttpLink`", () => {
-      fetchMock.postOnce(uri, response);
-      const client = new ApolloClient({
-        link: new BaseHttpLink({
-          uri,
-        }),
-        cache: new InMemoryCache(),
-        clientAwareness: {
-          name: "test-client",
-          version: "1.0.0",
-        },
-      });
-
-      void client.query({ query });
-      const headers = fetchMock.lastCall()![1]?.headers;
-      expect(headers).toStrictEqual({
-        accept: "application/graphql-response+json,application/json;q=0.9",
-        "content-type": "application/json",
-      });
-    });
-
-    test("`HttpLink` options have priotity over `ApolloClient` options", () => {
-      fetchMock.postOnce(uri, response);
-      const client = new ApolloClient({
-        link: new HttpLink({
-          uri,
-          clientAwareness: {
-            name: "overridden-client",
-            version: "2.0.0",
-          },
-        }),
-        cache: new InMemoryCache(),
-
-        clientAwareness: {
-          name: "test-client",
-          version: "1.0.0",
-        },
-      });
-
-      void client.query({ query });
-      const headers = fetchMock.lastCall()![1]?.headers;
-      expect(headers).toStrictEqual({
-        accept: "application/graphql-response+json,application/json;q=0.9",
-        "content-type": "application/json",
-        "apollographql-client-name": "overridden-client",
-        "apollographql-client-version": "2.0.0",
-      });
-    });
-  });
-
-  describe("BatchHttpLink integration", () => {
-    test("is part of `BatchHttpLink`", async () => {
-      fetchMock.postOnce(uri, [response, response]);
-      const client = new ApolloClient({
-        link: new BatchHttpLink({
-          uri,
-        }),
-        cache: new InMemoryCache(),
-        clientAwareness: {
-          name: "test-client",
-          version: "1.0.0",
-        },
-      });
-
-      void client.query({ query });
-      void client.query({ query, context: { queryDeduplication: false } });
-      await wait(10);
-
-      const headers = fetchMock.lastCall()![1]?.headers;
-      expect(headers).toStrictEqual({
-        accept: "application/graphql-response+json,application/json;q=0.9",
-        "content-type": "application/json",
-        "apollographql-client-name": "test-client",
-        "apollographql-client-version": "1.0.0",
-      });
-    });
-
-    test("is not part of `BaseBatchHttpLink`", async () => {
-      fetchMock.postOnce(uri, response);
-      const client = new ApolloClient({
-        link: new BaseBatchHttpLink({
-          uri,
-        }),
-        cache: new InMemoryCache(),
-        clientAwareness: {
-          name: "test-client",
-          version: "1.0.0",
-        },
-      });
-
-      void client.query({ query });
-      await wait(10);
-
-      const headers = fetchMock.lastCall()![1]?.headers;
-      expect(headers).toStrictEqual({
-        accept: "application/graphql-response+json,application/json;q=0.9",
-        "content-type": "application/json",
-      });
-    });
-
-    test("`BatchHttpLink` options have priority over `ApolloClient` options", async () => {
-      fetchMock.postOnce(uri, response);
-      const client = new ApolloClient({
-        link: new BatchHttpLink({
-          uri,
-
-          clientAwareness: {
-            name: "overridden-client",
-            version: "2.0.0",
-          },
-        }),
-        cache: new InMemoryCache(),
-
-        clientAwareness: {
-          name: "test-client",
-          version: "1.0.0",
-        },
-      });
-
-      void client.query({ query });
-      await wait(10);
-
-      const headers = fetchMock.lastCall()![1]?.headers;
-      expect(headers).toStrictEqual({
-        accept: "application/graphql-response+json,application/json;q=0.9",
-        "content-type": "application/json",
-        "apollographql-client-name": "overridden-client",
-        "apollographql-client-version": "2.0.0",
-      });
-    });
-
-    test("will batch requests with equal options", async () => {
-      fetchMock.postOnce(uri, [response, response]);
-      const client = new ApolloClient({
-        link: new BatchHttpLink({
-          uri,
-        }),
-        cache: new InMemoryCache(),
-
-        clientAwareness: {
-          name: "test-client",
-          version: "1.0.0",
-        },
-      });
-
-      void client.query({ query, context: { queryDeduplication: false } });
-      void client.query({ query, context: { queryDeduplication: false } });
-      await wait(10);
-
-      const headers = fetchMock.lastCall()![1]?.headers;
-      expect(headers).toStrictEqual({
-        accept: "application/graphql-response+json,application/json;q=0.9",
-        "content-type": "application/json",
-        "apollographql-client-name": "test-client",
-        "apollographql-client-version": "1.0.0",
-      });
-    });
-
-    test("will not batch requests with different options", async () => {
-      fetchMock.post(uri, [response], { repeat: 2 });
-      const client = new ApolloClient({
-        link: new BatchHttpLink({
-          uri,
-        }),
-        cache: new InMemoryCache(),
-
-        clientAwareness: {
-          name: "test-client",
-          version: "1.0.0",
-        },
-      });
-
-      void client.query({ query, context: { queryDeduplication: false } });
-      void client.query({
-        query,
-        context: {
-          queryDeduplication: false,
-          clientAwareness: {
-            name: "overridden-client",
-            version: "2.0.0",
-          },
-        },
-      });
-      await wait(10);
-
-      expect(fetchMock.calls()[0][1]?.headers).toStrictEqual({
-        accept: "application/graphql-response+json,application/json;q=0.9",
-        "content-type": "application/json",
-        "apollographql-client-name": "test-client",
-        "apollographql-client-version": "1.0.0",
-      });
-      expect(fetchMock.calls()[1][1]?.headers).toStrictEqual({
-        accept: "application/graphql-response+json,application/json;q=0.9",
-        "content-type": "application/json",
-        "apollographql-client-name": "overridden-client",
-        "apollographql-client-version": "2.0.0",
-      });
-    });
-  });
 });
 
 describe("feature: enhanced client awareness", () => {
@@ -505,169 +264,5 @@ describe("feature: enhanced client awareness", () => {
 
     void client.query({ query });
     expect(terminatingLink.operation?.extensions).toStrictEqual({});
-  });
-
-  describe("HttpLink integration", () => {
-    test("is part of `HttpLink`", () => {
-      fetchMock.postOnce(uri, response);
-      const client = new ApolloClient({
-        link: new HttpLink({
-          uri,
-        }),
-        cache: new InMemoryCache(),
-      });
-
-      void client.query({ query });
-      const body = JSON.parse(fetchMock.lastCall()![1]?.body as string);
-      expect(body.extensions).toStrictEqual({
-        clientLibrary: {
-          name: "@apollo/client",
-          version,
-        },
-      });
-    });
-
-    test("is not part of `BaseHttpLink`", () => {
-      fetchMock.postOnce(uri, response);
-      const client = new ApolloClient({
-        link: new BaseHttpLink({
-          uri,
-        }),
-        cache: new InMemoryCache(),
-      });
-
-      void client.query({ query });
-      const body = JSON.parse(fetchMock.lastCall()![1]?.body as string);
-      expect(body.extensions).not.toBeDefined();
-    });
-
-    test("can be disabled from `HttpLink`", () => {
-      fetchMock.postOnce(uri, response);
-      const client = new ApolloClient({
-        link: new HttpLink({
-          uri,
-          enhancedClientAwareness: { transport: false },
-        }),
-        cache: new InMemoryCache(),
-      });
-
-      void client.query({ query });
-      const body = JSON.parse(fetchMock.lastCall()![1]?.body as string);
-      expect(body.extensions).not.toBeDefined();
-    });
-
-    test("can also be disabled by disabling `includeExtensions`", () => {
-      fetchMock.postOnce(uri, response);
-      const client = new ApolloClient({
-        link: new HttpLink({
-          uri,
-          includeExtensions: false,
-        }),
-        cache: new InMemoryCache(),
-      });
-
-      void client.query({ query });
-      const body = JSON.parse(fetchMock.lastCall()![1]?.body as string);
-      expect(body.extensions).not.toBeDefined();
-    });
-  });
-
-  describe("BatchHttpLink integration", () => {
-    test("is part of `BatchHttpLink`", async () => {
-      fetchMock.postOnce(uri, [response, response]);
-      const client = new ApolloClient({
-        link: new BatchHttpLink({
-          uri,
-        }),
-        cache: new InMemoryCache(),
-      });
-
-      void client.query({ query });
-      void client.query({ query, context: { queryDeduplication: false } });
-      await wait(10);
-
-      const body = JSON.parse(fetchMock.lastCall()![1]?.body as string);
-      expect(body[0].extensions).toStrictEqual({
-        clientLibrary: {
-          name: "@apollo/client",
-          version,
-        },
-      });
-      expect(body[1].extensions).toStrictEqual({
-        clientLibrary: {
-          name: "@apollo/client",
-          version,
-        },
-      });
-    });
-
-    test("is not part of `BaseBatchHttpLink`", async () => {
-      fetchMock.postOnce(uri, [response, response]);
-      const client = new ApolloClient({
-        link: new BaseBatchHttpLink({
-          uri,
-        }),
-        cache: new InMemoryCache(),
-      });
-
-      void client.query({ query });
-      void client.query({ query, context: { queryDeduplication: false } });
-      await wait(10);
-
-      const body = JSON.parse(fetchMock.lastCall()![1]?.body as string);
-      expect(body[0].extensions).not.toBeDefined();
-      expect(body[1].extensions).not.toBeDefined();
-    });
-
-    test("can be disabled by disabling `includeExtensions`", async () => {
-      fetchMock.postOnce(uri, [response, response]);
-      const client = new ApolloClient({
-        link: new BatchHttpLink({
-          uri,
-          includeExtensions: false,
-        }),
-        cache: new InMemoryCache(),
-      });
-
-      void client.query({ query });
-      void client.query({ query, context: { queryDeduplication: false } });
-      await wait(10);
-
-      const body = JSON.parse(fetchMock.lastCall()![1]?.body as string);
-      expect(body[0].extensions).not.toBeDefined();
-      expect(body[1].extensions).not.toBeDefined();
-    });
-
-    test("can send mixed requests with ECA enabled and disabled", async () => {
-      fetchMock.postOnce(uri, [response, response]);
-      const client = new ApolloClient({
-        link: ApolloLink.split(
-          (operation) => operation.variables.a === 1,
-          new ClientAwarenessLink({
-            enhancedClientAwareness: { transport: "extensions" },
-          }),
-          new ClientAwarenessLink({
-            enhancedClientAwareness: { transport: false },
-          })
-        ).concat(new BaseBatchHttpLink({ uri })),
-        cache: new InMemoryCache(),
-      });
-
-      void client.query({ query, variables: { a: 1 } });
-      void client.query({
-        query,
-        variables: { a: 2 },
-      });
-      await wait(10);
-
-      const body = JSON.parse(fetchMock.lastCall()![1]?.body as string);
-      expect(body[0].extensions).toStrictEqual({
-        clientLibrary: {
-          name: "@apollo/client",
-          version,
-        },
-      });
-      expect(body[1].extensions).not.toBeDefined();
-    });
   });
 });
